Added fake overloads to check for strict signatures.
Specifically, the fake overloads accept any set of arguments but are themselves
deleted, so any non-exact match to the intended functions will instead match to
the fake overload and trigger a compile-time error. In this manner, we can
prevent implicit casts that might cause overflows.

We also refactor for some minor improvements:

- Overhaul the test_*_access() functions to avoid allocating storage
  within the loop body, for a slight efficiency improvement.
- Slap const on function arguments and variables, for safety.
- Avoid hard-coded size_t's, use I<decltype()> to infer the type.